refactor precompile, particularly around LSigs
the main goal here is to remove the implicit interlink between Python/PyTeal/Beaker and how we interact with these applications containing precompiles. This doesn't fully solve for that, the embedded data will need to be in the app-spec most likely, but it does remove an implicit linking of the Python code contained within an Application and getting an lsig signer.